Refactor search game
Previously I was using a lib to convert XML to elixir map (because I was
too lazy to really learn how to deal with XML), but I was feeling it too
hackey for my taste. For the game/thing endpoint I decided to use a real
XML tool, then for consistency I refactored the search function and
removed the elixir_xml_to_map dependency.